+++ openproblems/tasks/spatial_decomposition/methods/cell2location.py
@@ -1,7 +1,7 @@
 from ....tools.decorators import method
 from ....tools.normalize import log_cpm
 from ....tools.utils import check_version
-from .._utils import split_sc_and_sp
+from ..utils import split_sc_and_sp
 
 # import cell2location
