+++ openproblems/tasks/spatial_decomposition/methods/nmfreg.py
@@ -1,6 +1,6 @@
 from ....tools.decorators import method
 from ....tools.utils import check_version
-from .._utils import split_sc_and_sp
+from ..utils import split_sc_and_sp
 
 import numpy as np
