+++ openproblems/tasks/spatial_decomposition/methods/vanillanmf.py
@@ -1,7 +1,7 @@
 from ....tools.decorators import method
 from ....tools.utils import check_version
-from .._utils import obs_means
-from .._utils import split_sc_and_sp
+from ..utils import obs_means
+from ..utils import split_sc_and_sp
 
 
 @method(
