+++ openproblems/tasks/spatial_decomposition/utils.py
